+++ src/graphdatascience/procedure_surface/cypher/topological_link_prediction_cypher_endpoints.py
@@ -7,6 +7,7 @@ from neo4j.graph import Node
 from graphdatascience.error.cypher_warning_handler import filter_id_func_deprecation_warning
 from graphdatascience.procedure_surface.api.topological_link_prediction_endpoints import (
     Direction,
+    DirectionLike,
     TopologicalLinkPredictionEndpoints,
 )
 from graphdatascience.query_runner.query_mode import QueryMode
@@ -24,7 +25,7 @@ class TopologicalLinkPredictionCypherEndpoints(TopologicalLinkPredictionEndpoint
         node1: int | Node,
         node2: int | Node,
         relationship_query: str | None = None,
-        direction: Direction = Direction.BOTH,
+        direction: DirectionLike = Direction.BOTH,
     ) -> float:
         return self._run_similarity_function(
             "gds.linkprediction.adamicAdar", node1, node2, relationship_query, direction
@@ -36,7 +37,7 @@ class TopologicalLinkPredictionCypherEndpoints(TopologicalLinkPredictionEndpoint
         node1: int | Node,
         node2: int | Node,
         relationship_query: str | None = None,
-        direction: Direction = Direction.BOTH,
+        direction: DirectionLike = Direction.BOTH,
     ) -> float:
         return self._run_similarity_function(
             "gds.linkprediction.commonNeighbors", node1, node2, relationship_query, direction
@@ -48,7 +49,7 @@ class TopologicalLinkPredictionCypherEndpoints(TopologicalLinkPredictionEndpoint
         node1: int | Node,
         node2: int | Node,
         relationship_query: str | None = None,
-        direction: Direction = Direction.BOTH,
+        direction: DirectionLike = Direction.BOTH,
     ) -> float:
         return self._run_similarity_function(
             "gds.linkprediction.preferentialAttachment", node1, node2, relationship_query, direction
@@ -60,7 +61,7 @@ class TopologicalLinkPredictionCypherEndpoints(TopologicalLinkPredictionEndpoint
         node1: int | Node,
         node2: int | Node,
         relationship_query: str | None = None,
-        direction: Direction = Direction.BOTH,
+        direction: DirectionLike = Direction.BOTH,
     ) -> float:
         return self._run_similarity_function(
             "gds.linkprediction.resourceAllocation", node1, node2, relationship_query, direction
@@ -92,7 +93,7 @@ class TopologicalLinkPredictionCypherEndpoints(TopologicalLinkPredictionEndpoint
         node1: int | Node,
         node2: int | Node,
         relationship_query: str | None = None,
-        direction: Direction = Direction.BOTH,
+        direction: DirectionLike = Direction.BOTH,
     ) -> float:
         return self._run_similarity_function(
             "gds.linkprediction.totalNeighbors", node1, node2, relationship_query, direction
@@ -104,13 +105,15 @@ class TopologicalLinkPredictionCypherEndpoints(TopologicalLinkPredictionEndpoint
         node1: int | Node,
         node2: int | Node,
         relationship_query: str | None,
-        direction: Direction,
+        direction: DirectionLike,
     ) -> float:
         query = f"""
         MATCH (n), (m)
         WHERE id(n) = $node1 AND id(m) = $node2
         RETURN {function}(n, m, $config) AS score
         """
+
+        direction = Direction.of(direction)
 
         config = {
             "direction": direction.value,