+++ src/graphdatascience/query_runner/__init__.py
@@ -1,5 +1,5 @@
-from .query_mode import QueryMode
+from .query_mode import QueryMode, QueryModeLike
 from .query_runner import QueryRunner
 from .query_type import QueryType
 
-__all__ = ["QueryRunner", "QueryMode", "QueryType"]
+__all__ = ["QueryRunner", "QueryMode", "QueryModeLike", "QueryType"]